feat: implement profile page that renders favorite news, news can be deleted from favorite here as well as on news page, split news card content to different file; fix eslint warnings
