+++ react-spa-test/src/components/News/index.scss
@@ -6,70 +6,6 @@
   justify-content: center;
 }
 
-.news-card {
-  & .news-favorite-icon {
-    position: absolute;
-    top: 0;
-    right: 15px;
-  }
-  & .filled-icon {
-    color: $favorite-icon;
-  }
-}
-
-.news-card-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  width: 350px;
-  margin: 10px;
-}
-
-.news-card-title {
-  h3 {
-    margin-top: 20px;
-    font-size: 20px;
-    text-decoration: underline;
-    &:hover {
-      color: $theme-color2
-    }
-  }
-}
-
-.image-container {
-  position: relative;
-  margin: auto;
-  max-width: fit-content;
-}
-
-.news-image {
-  max-width: 100%;
-  height: 184px;
-}
-
-.source-info {
-  display: flex;
-  justify-content: space-between;
-  font-size: 10px;
-  font-weight: 500;
-  color: $light-text;
-  letter-spacing: 1px;
-  & p {
-    margin-top: 10px;
-    margin-bottom: 3px;
-  }
-}
-
-.news-description {
-  text-align: justify;
-}
-
-.news-divider {
-  font-weight: 800;
-  font-size: 20px;
-  color: $theme-color2;
-}
-
 .backtop-button {
   height: 40px;
   width: 40px;
@@ -80,24 +16,3 @@
   text-align: center;
   font-size: 14px;
 };
-
-@media(max-width: 575px) {
-  .news-card-container {
-    margin: 6px;
-    width: 235px;
-  }
-
-  .news-image {
-    height: 124px;
-  }
-
-  .news-card-title {
-    h3 {
-      overflow: hidden;
-      text-overflow: ellipsis;
-      white-space: nowrap;
-      font-size: 17px;
-      line-height: 23px;
-    }
-  }
-}
