+++ react-spa-test/src/components/NewsCard/index.scss
@@ -1,0 +1,86 @@
+@import '../../@constants/constants';
+
+.news-card {
+  & .news-favorite-icon {
+    position: absolute;
+    top: 0;
+    right: 15px;
+  }
+  & .filled-icon {
+    color: $favorite-icon;
+  }
+}
+
+.news-card-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  width: 350px;
+  margin: 10px;
+}
+
+.news-card-title {
+  h3 {
+    margin-top: 20px;
+    font-size: 20px;
+    text-decoration: underline;
+    &:hover {
+      color: $theme-color2
+    }
+  }
+}
+
+.image-container {
+  position: relative;
+  margin: auto;
+  max-width: fit-content;
+}
+
+.news-image {
+  max-width: 100%;
+  height: 184px;
+}
+
+.source-info {
+  display: flex;
+  justify-content: space-between;
+  font-size: 10px;
+  font-weight: 500;
+  color: $light-text;
+  letter-spacing: 1px;
+  & p {
+    margin-top: 10px;
+    margin-bottom: 3px;
+  }
+}
+
+.news-description {
+  text-align: justify;
+}
+
+.news-divider {
+  font-weight: 800;
+  font-size: 20px;
+  color: $theme-color2;
+}
+
+@media(max-width: 575px) {
+  .news-card-container {
+    margin: 6px;
+    width: 235px;
+  }
+
+  .news-image {
+    height: 124px;
+  }
+
+  .news-card-title {
+    h3 {
+      overflow: hidden;
+      text-overflow: ellipsis;
+      white-space: nowrap;
+      font-size: 17px;
+      line-height: 23px;
+    }
+  }
+}
