+++ react-spa-test/src/components/NewsCard/models.ts
@@ -1,0 +1,5 @@
+import { NewsItem } from 'reducers/news/models';
+
+export default interface NewsCardProps {
+  item: NewsItem;
+}
