+++ react-spa-test/src/components/Profile/index.scss
@@ -1,0 +1,5 @@
+.profile {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+}
