+++ react-spa-test/src/reducers/authorization/index.ts
@@ -1,14 +1,14 @@
-import { createSlice } from '@reduxjs/toolkit';
+import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
 const currentState: string | null = localStorage.getItem('jwt-token');
 const initialState: boolean =
-  currentState === null ? false : JSON.parse(currentState);
+  currentState === null ? false : (JSON.parse(currentState) as boolean);
 
 const authorizationSlice = createSlice({
   name: 'authorization',
   initialState,
   reducers: {
-    setIsAuth: (state, { payload }) => payload,
+    setIsAuth: (state, { payload }: PayloadAction<boolean>) => payload,
   },
 });
