+++ react-spa-test/src/reducers/authorization/models.ts
@@ -1,1 +1,1 @@
-export {}
+export {};
