+++ react-spa-test/src/reducers/loginForm/index.ts
@@ -1,4 +1,4 @@
-import { createSlice } from '@reduxjs/toolkit';
+import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import LoginForm from './models';
 
 const initialState: LoginForm = { isLoginFormOpen: false };
@@ -7,7 +7,7 @@ const loginFormSlice = createSlice({
   name: 'logIn',
   initialState,
   reducers: {
-    toggleLoginForm: (state, { payload }) => {
+    toggleLoginForm: (state, { payload }: PayloadAction<LoginForm>) => {
       state.isLoginFormOpen = payload.isLoginFormOpen;
     },
   },
